started documenting errors
